% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convertir_a_factor.R
\name{convertir_a_factor}
\alias{convertir_a_factor}
\title{Convertir un conjunto de variables a factor.}
\usage{
convertir_a_factor(data, columnas_a_convertir)
}
\arguments{
\item{data}{Archivo de datos.}

\item{columnas_a_convertir}{Nombre de las columnas que desea pasar a factor.}
}
\value{
Archivo de datos, con las variables seleccionadas transformadas a factor.
}
\description{
Convertir un conjunto de variables a factor.
}
\examples{
convertir_a_factor("Nombre_conjunto_de_datos", c("columna1","columna2","columna_n"))

}
