% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/McIntosh-package.R
\docType{package}
\name{McIntosh-package}
\alias{McIntosh}
\alias{McIntosh-package}
\title{McIntosh: Análisis combinado - Efectos fijos}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Fernando da Fonseca \email{fernandodafonseca@mdp.edu.ar} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
